+++ src/chat/ChatHistory.jsx
@@ -4,7 +4,7 @@
 import React from "react";
 import { createPortal } from "react-dom";
 import { Icon } from "../components/Icon.jsx";
-import { usePortalPopover } from "./usePortalPopover.js";
+import { usePortalPopover } from "../hooks/usePortalPopover.js";
 import { ChatThemePicker } from "./ChatThemePicker.jsx";
 
 function ChatHistory({ convos, activeId, onPick, onDelete, onOpen, loading, showThemePicker }) {